Brings snoups/remnashop dev up to Aug 2026: grace mode for expired
subscriptions, atomic trial activation, panel user reactivation on paid
renewal, payment gateway fixes (UrlPay/MulenPay/RollyPay), password
reset rate limiting, unified expiration webhook. Adapted to remnapy
3.2.1 (id-based users, apply_grace uses user_id); migration renumbered
to 0047 after upstream 0041-0046.

+++ src/core/utils/converters.py
@@ -69,7 +69,7 @@ def gb_to_bytes(value: Optional[int]) -> int:
     return _round_decimal(Decimal(value) * _GB_FACTOR)
 
 
-def bytes_to_gb(value: Optional[int]) -> int:
+def bytes_to_gb(value: Optional[float]) -> int:
     if not value:
         return 0
